FILE: moves Places class to internal/ folder within lib
