+++ libs/globetrotter/learn/data-access/src/lib/quiz.service.spec.ts
@@ -4,9 +4,9 @@ import { of } from 'rxjs';
 
 import { Country } from '@atocha/globetrotter/shared/util';
 import { QuizType, shuffle } from '@atocha/globetrotter/learn/util';
+import { Places } from './internal/places';
 import { QuizService } from './quiz.service';
 import { PlaceService } from './place.service';
-import { Places } from './places';
 
 jest.mock('@atocha/globetrotter/learn/util', () => ({
   QuizType: {
